Save importing fixes, Elite Four Match Call tweak
Elite Four will only be added to Match Call after beating them in a rematch. This is so the player can't call them and have them welcoming back the player to the Pokemon League in case it's still being overhauled (haven't beaten Juan in the post-game and done the other stuff).

Save fixes:

*Remove Elite Four/Champion Match Calls for first-time imported saves (version check var won't be enabled until ready for some form of release).
*Remove registered item. Mainly for multi-register menu when its issues gets fully worked out.
*Swap player's bike with the Bicycle. Only happens if player already has a bike.
*Reset Juan's battle state and remove him from Match Call if player has the Rain Badge and hasn't beaten the game.
*Fix game not showing Mystery Gift man to give Aurora Ticket if player is already Champion in imported save file.