Restore blanked wallpaper symbols from JPN version
The Cross, Bolt, and Plusle icons, as well as the ribbon in the Ribbon "Friends" wallpaper were removed from international versions of Emerald, likely for religious reasons and for potential use of Red Cross imagery.

IDK how the icons were saved for the decomp, but their palette order is reversed from what it actually is in-game.