Battle Pike double battle fainted Pokemon fix
Prevent hinted room from being the double battle room if player has only one non-fainted Pokemon. The hinted room is set upon receiving a hint, but there isn't an extra check if a Pokemon faints from poision between getting the hint and the hinted room loading.

Note that Poison on the overworld is intentionally enabled here and the Battle Pyramid in the hack, but nowhere else.