Add one-time TMs to player's bag in imported saves
Checks if the player already got the TM based on assigned flags (or variable, in the case of TM12). This is in case the imported save already used or got rid of the TMs in question.

This fixes the player potentially not being able to use Secret Power in imported saves.