Swap enemy Lanturn's illegal Earthquake with Surf
Ace Trainers Dianne and Felix are set up to encourage a double battle. Given the similarities in how their rosters are set (Claydol that knows Skill Swap and Earthquake, a second Pokemon that has a strong STAB special move), it seems likely that Lanturn was only meant to know Thunderbolt and nothing else. I decided to switch it out with a more-fitting move rather than leave it at one move altogether in order to retain the level of difficulty.