Add Hyper Training to Battle Frontier ESC
A Hyper Training expert now appears near the PC at the Exchange Service Corner. One or all available stats can be trained for Lv. 50+ Pokemon for 24 BP each (value based on weighted equivalent to buying Bottle Caps in Gen VIII). Dialogue is inspired by Hyper Training NPCs in the official games.

Hyper Trained stats appear on the summary screen with gold text. The summary screen palette was updated to include an extra shade of gold, which has also been applied to stars for stats with an IV of 31.